refactor: DTO 패키지 정리
