+++ src/main/java/wooteco/subway/domain/path/PathFindingStrategy.java
@@ -1,6 +1,9 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 import java.util.List;
+
+import wooteco.subway.domain.Lines;
+import wooteco.subway.domain.Station;
 
 public interface PathFindingStrategy {
     int getShortestDistance(Station source, Station target, Lines lines);
