+++ src/main/java/wooteco/subway/domain/path/WeightEdgeWithLineId.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
