+++ src/main/java/wooteco/subway/dto/controller/request/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.request;
+package wooteco.subway.dto.controller.request;
 
 public class SectionRequest {
     private Long upStationId;
