+++ src/main/java/wooteco/subway/dto/controller/request/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.request;
+package wooteco.subway.dto.controller.request;
 
 public class StationRequest {
     private String name;
