+++ src/main/java/wooteco/subway/dto/service/request/LineServiceRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service.request;
 
 public class LineServiceRequest {
     private final String name;
