+++ src/main/java/wooteco/subway/dto/service/request/SectionCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.info;
+package wooteco.subway.dto.service.request;
 
 public class SectionCreateRequest {
     private final long lineId;
