+++ src/main/java/wooteco/subway/service/StationService.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.info.StationDto;
+import wooteco.subway.dto.service.StationDto;
 
 @Service
 public class StationService {
