+++ src/test/java/wooteco/subway/acceptance/LineAcceptanceTest.java
@@ -14,7 +14,7 @@ import org.springframework.http.HttpStatus;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import wooteco.subway.dto.response.LineResponse;
+import wooteco.subway.dto.controller.response.LineResponse;
 
 @DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
