+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -14,7 +14,7 @@ import org.springframework.http.HttpStatus;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import wooteco.subway.dto.response.StationResponse;
+import wooteco.subway.dto.controller.response.StationResponse;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
