+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.info.LineServiceRequest;
-import wooteco.subway.dto.info.LineServiceResponse;
-import wooteco.subway.dto.info.LineUpdateRequest;
+import wooteco.subway.dto.service.request.LineServiceRequest;
+import wooteco.subway.dto.service.request.LineUpdateRequest;
+import wooteco.subway.dto.service.response.LineServiceResponse;
 
 public class LineServiceTest {
