+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.info.StationDto;
+import wooteco.subway.dto.service.StationDto;
 
 public class StationServiceTest {
     private StationService stationService;
